불필요한 코드 제거 및 변경
1. 파일 업로드 삭제 부분은 먼저 데이터를 삭제 및 생성 후 파일을 맨 마지막 부분에서 실행

1-1. 하지만 thumbnail 같이 중간 부분에서 이미지를 조작해야하는 경우는 제외

2. 컨트롤러 공용 메소드 변경 및 삭제

+++ backend/src/controller/user.admin.controller.ts
@@ -1,6 +1,6 @@
 import dayjs from "dayjs";
 import randomString from "randomstring";
-import { Op, OrderItem, WhereOptions } from "sequelize";
+import { Op, OrderItem, Transaction, WhereOptions } from "sequelize";
 import { File } from "formidable";
 import { boolean } from "boolean";
 
@@ -12,7 +12,6 @@ import sequelize from "../model";
 import { User, IUserResponseWithCount, PageOptions, SearchOptions, FilterOptions, IUpdateWithAdmin, ICreateWithAdmin } from "../model/user.model";
 import { Solution } from "../model/solution.model";
 import { UserRole } from "../model/userRole.model";
-import { ErrorImage } from "../model/errorImage.model";
 import { Couple } from "../model/couple.model";
 import { Album } from "../model/album.model";
 import { Inquire } from "../model/inquire.model";
@@ -86,7 +85,7 @@ const createCode = async (): Promise<string> => {
     return code;
 };
 
-const uploadProfile = async (userId: number, file: File) => {
+const createProfilePath = (userId: number, file: File): string | null => {
     let path: string | null = "";
     const reqFileName = file.originalFilename!;
     const isDefault = isDefaultFile(reqFileName);
@@ -95,12 +94,8 @@ const uploadProfile = async (userId: number, file: File) => {
      * Frontend에선 static으로 default.jpg,png,svg 셋 중 하나 갖고있다가
      * 사용자가 profile을 내리면 그걸로 넣고 요청
      */
-    if (isDefault) {
-        path = null;
-    } else {
-        path = `${FOLDER_NAME}/${userId}/profile/${dayjs().valueOf()}.${reqFileName}`;
-        await uploadFile(path, file.filepath);
-    }
+    if (isDefault) path = null;
+    else path = `${FOLDER_NAME}/${userId}/profile/${dayjs().valueOf()}.${reqFileName}`;
 
     return path;
 };
@@ -192,11 +187,13 @@ const controller = {
             data.code = await createCode();
         }
 
-        const transaction = await sequelize.transaction();
+        let transaction: Transaction | undefined = undefined;
         const hash: string = await createDigest(data.password);
         data.password = hash;
 
         try {
+            transaction = await await sequelize.transaction();
+
             const createdUser: User = await User.create(
                 {
                     snsId: data.snsId,
@@ -222,20 +219,22 @@ const controller = {
             );
 
             if (file) {
-                data.profile = await uploadProfile(createdUser.userId, file);
-                if (data.profile) isUpload = true;
+                data.profile = createProfilePath(createdUser.userId, file);
+                await createdUser.update({ profile: data.profile }, { transaction });
+                await uploadFile(path, file.filepath);
+                isUpload = true;
             }
 
-            await createdUser.update({ profile: data.profile }, { transaction });
-            transaction.commit();
+            await transaction.commit();
         } catch (error) {
-            transaction.rollback();
-
             // Firebase에는 업로드 되었지만 DB 오류가 발생했다면 Firebase Profile 삭제
             if (file && isUpload) {
                 await deleteFile(path!);
                 logger.error(`After updating the firebase, a db error occurred and the firebase profile is deleted => ${path}`);
             }
+
+            if (transaction) await transaction.rollback();
+            logger.error(`User create error => ${JSON.stringify(error)}`);
 
             throw error;
         }
@@ -277,45 +276,51 @@ const controller = {
             where: { userId }
         });
         if (!user) throw new NotFoundError("Not Found User");
-        let prevProfile: string | null = user.profile;
 
-        const transaction = await sequelize.transaction();
+        let prevProfile: string | null = user.profile;
+        let transaction: Transaction | undefined = undefined;
 
         try {
+            transaction = await sequelize.transaction();
+
             if (data.role) {
                 const userRole: UserRole | null = await UserRole.findOne({ where: { userId } });
 
                 if (!userRole) throw new NotFoundError(`Not found User Role. userId : ${userId}`);
                 await userRole.update({ roleId: data.role }, { transaction });
             }
-            if (file) {
-                data.profile = await uploadProfile(userId, file);
-                if (data.profile) isUpload = true;
-                else if (prevProfile && !data.profile) await deleteFile(prevProfile);
-            }
+
             if (data.password) data.password = await createDigest(data.password);
             if (boolean(data.deleted)) data.deletedTime = new Date(dayjs().valueOf());
             else if (data.deleted !== undefined && boolean(data.deleted) === false) data.deletedTime = null;
 
+            if (file) {
+                data.profile = createProfilePath(userId, file);
+
+                // profile 있으면 업로드
+                if (data.profile) {
+                    await uploadFile(data.profile, file.filepath);
+                    isUpload = true;
+
+                    if (prevProfile) await deleteFile(prevProfile); // 전에 있던 profile 삭제
+                } else if (prevProfile && !data.profile) {
+                    // default 이미지로 변경시
+                    await deleteFile(prevProfile);
+                }
+            }
+
             await user.update(data, { transaction });
             logger.debug(`Update Data => ${JSON.stringify(data)}`);
 
-            // 이미 profile이 있다면 Firebase에서 삭제
-            if (prevProfile && data.profile) {
-                await deleteFile(prevProfile);
-                logger.debug(`Deleted Previous Profile => ${prevProfile}`);
-            }
-
-            transaction.commit();
+            await transaction.commit();
         } catch (error) {
-            transaction.rollback();
-
             // Firebase에는 업로드 되었지만 DB 오류가 발생했다면 Firebase Profile 삭제
             if (data.profile && isUpload) {
                 await deleteFile(path!);
                 logger.error(`After updating the firebase, a db error occurred and the firebase profile is deleted => ${path}`);
             }
 
+            if (transaction) await transaction.rollback();
             throw error;
         }
     },
@@ -353,34 +358,40 @@ const controller = {
         const userHasCouple = users.filter((user: User) => {
             if (user.cupId) return user;
         });
+        let transaction: Transaction | undefined = undefined;
 
-        // Inquire 삭제
-        userHasInquiry.forEach(async (user: User) => {
-            user.inquires!.forEach(async (inquire: Inquire) => {
-                await inquireController.deleteInquire(inquire.inquireId);
-                // soluton image 삭제
-                // if (inquire.solution) await solutionController.deleteSolution(inquire.solution.solutionId);
-            });
-        });
+        try {
+            transaction = await sequelize.transaction();
 
-        // Album 삭제 및 Couple 삭제
-        userHasCouple.forEach(async (user: User) => {
-            const otherUser: User[] = (await user.couple!.getUsers()).filter((coupleUser: User) => {
-                if (coupleUser.userId != user.userId) return user;
-            });
-            const albums: Album[] = await user.couple!.getAlbums();
+            // inquire 삭제(image), couple 삭제(thumbnail), 다른 couple user cupId null 처리, user thumbnail 삭제, 유저 삭제
 
-            if (albums) {
-                albums.forEach(async (album: Album) => {
-                    await albumController.deleteAlbum(user.cupId!, album.albumId);
-                });
+            // Inquire 삭제
+            for (const user of userHasInquiry) {
+                for (const inquire of user.inquires!) {
+                    await inquireController.deleteInquire(inquire.inquireId);
+                    // soluton image 삭제
+                    // if (inquire.solution) await solutionController.deleteSolution(inquire.solution.solutionId);
+                }
             }
 
-            await otherUser[0].update({ cupId: null });
+            for (const user of userHasCouple) {
+                const otherUser: User[] = (await user.couple!.getUsers()).filter((coupleUser: User) => {
+                    if (coupleUser.userId != user.userId) return user;
+                });
+                const albums: Album[] = await user.couple!.getAlbums();
 
-            if (user.couple!.thumbnail) await deleteFile(user.couple!.thumbnail);
-            await user.couple!.destroy();
-        });
+                if (albums) {
+                    for (const album of albums) {
+                        await albumController.deleteAlbum(user.cupId!, album.albumId);
+                    }
+                }
+
+                await otherUser[0].update({ cupId: null });
+
+                if (user.couple!.thumbnail) await deleteFile(user.couple!.thumbnail);
+                await user.couple!.destroy();
+            }
+        } catch (error) {}
 
         users.forEach(async (user: User) => {
             const profile: string | null = user.profile;
